Update styles, Add separate styles.js file to import customStyles into Select input components
