+++ src/components/Input/Input.css
@@ -1,18 +1,17 @@
+.select-img-container {
+  display: flex;
+  flex-wrap: wrap;
+}
+
 .amount-input {
   height: 4.3rem;
   margin-right: 1.2rem;
   width: 20rem;
   border-radius: 18px;
   border: 1px solid #463182;
+  font-family: 'Nunito', sans-serif;
+  font-size: 1.8rem;
 }
-
-/* .select { */
-/* height: 4.3rem; */
-/* margin-right: 1.2rem;
-  width: 20rem;
-  border-radius: 18px; */
-/* border: 1px solid #463182; */
-/* } */
 
 .arrows-image {
   width: 3rem;
@@ -22,4 +21,8 @@
   color: #fff;
   margin-bottom: 0.3rem;
   font-size: 1.1rem;
+}
+
+.image-container {
+  margin-top: auto;
 }
